# contracts/openapi/endpoint-platform-v1.yaml
openapi: "3.1.0"
info:
  title: "Endpoint Platform API"
  version: "v1"
paths:
...
  "/api/v1/devices/{device_id}/operations":
    post:
      tags:
        - "endpoint-operations"
      summary: "Create Device Operation"
      description: "Create or replay one operation owned by the authenticated service client."
      operationId: "create_device_operation_api_v1_devices__device_id__operations_post"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "Idempotency-Key"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 8
            maxLength: 128
            pattern: "^[!-~][ -~]{6,126}[!-~]$"
            title: "Idempotency-Key"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/EndpointOperationCreateV1"
      responses:
        "201":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "200":
          description: "Replayed endpoint operation"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
        "409":
          description: "Idempotency key conflict"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "503":
          description: "Safe result unavailable"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.create"
  "/api/v1/operations/{operation_id}":
    get:
      tags:
        - "endpoint-operations"
      summary: "Read Endpoint Operation"
      description: "Read one operation through stable service-client ownership."
      operationId: "read_endpoint_operation_api_v1_operations__operation_id__get"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "operation_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Operation Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Endpoint operation not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "503":
          description: "Safe result unavailable"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.read"
  "/api/v1/operations/{operation_id}/cancel":
    post:
      tags:
        - "endpoint-operations"
      summary: "Cancel Endpoint Operation"
      description: "Cancel one owner-scoped operation only while it remains undelivered."
      operationId: "cancel_endpoint_operation_api_v1_operations__operation_id__cancel_post"
      security:
        -
          ServiceBearer: []
      parameters:
        -
          name: "operation_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Operation Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OperationResponseEnvelope"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "404":
          description: "Endpoint operation not found"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "409":
          description: "Operation has already reached delivery or a terminal state"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "401":
          description: "Service authentication failed"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "403":
          description: "Service scope is insufficient"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
        "422":
          description: "Validation Error"
          headers:
            X-Correlation-ID:
              description: "Exact echo of the request tracing header."
              schema:
                type: "string"
                minLength: 1
                maxLength: 128
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      x-required-scopes:
        - "operations.cancel"
...
components:
  schemas:
...
    EndpointOperationCreateV1:
      additionalProperties: false
      description: "Strict service request for one bounded diagnostic collection."
      properties:
        schema_version:
          const: "endpoint_operation_create_v1"
          title: "Schema Version"
          type: "string"
        capability:
          const: "context.diagnostic.collect"
          title: "Capability"
          type: "string"
        parameters:
          "$ref": "#/components/schemas/EndpointOperationCreateV1__DiagnosticCollectionParametersV1"
      required:
        - "schema_version"
        - "capability"
        - "parameters"
      title: "EndpointOperationCreateV1"
      type: "object"
    EndpointOperationCreateV1__DiagnosticCollectionParametersV1:
      additionalProperties: false
      description: "The only public operation input supported in v1."
      properties:
        reason:
          maxLength: 256
          minLength: 1
          pattern: "^(?![\\s\\S]*://)[^\\x00-\\x1f\\x7f]{1,256}(?![\\s\\S])"
          title: "Reason"
          type: "string"
      required:
        - "reason"
      title: "DiagnosticCollectionParametersV1"
      type: "object"
...
    EndpointOperationV1:
      "$comment": "Model-only constraints: deadline_at must be after created_at; completed_at is set only for terminal states; an available safe result requires succeeded status and may expire."
      additionalProperties: false
      description: "Service-visible operation lifecycle without command or agent internals."
      properties:
        schema_version:
          const: "endpoint_operation_v1"
          title: "Schema Version"
          type: "string"
        operation_id:
          format: "uuid"
          title: "Operation Id"
          type: "string"
        device_id:
          format: "uuid"
          title: "Device Id"
          type: "string"
        capability:
          const: "context.diagnostic.collect"
          title: "Capability"
          type: "string"
        status:
          enum:
            - "queued"
            - "delivered"
            - "acknowledged"
            - "running"
            - "succeeded"
            - "failed"
            - "canceled"
            - "expired"
          title: "Status"
          type: "string"
        created_at:
          format: "date-time"
          title: "Created At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        deadline_at:
          format: "date-time"
          title: "Deadline At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        completed_at:
          anyOf:
            -
              format: "date-time"
              type: "string"
              "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
            -
              type: "null"
          default: null
          title: "Completed At"
        result_available:
          title: "Result Available"
          type: "boolean"
        warnings:
          items:
            enum:
              - "command_failed"
              - "command_timed_out"
              - "data_truncated"
              - "permission_denied"
              - "probe_unavailable"
              - "redaction_applied"
              - "source_unavailable"
              - "unsupported_platform"
            type: "string"
          maxItems: 16
          title: "Warnings"
          type: "array"
      required:
        - "schema_version"
        - "operation_id"
        - "device_id"
        - "capability"
        - "status"
        - "created_at"
        - "deadline_at"
        - "result_available"
      title: "EndpointOperationV1"
      type: "object"
    EndpointDiagnosticResultV1:
      additionalProperties: false
      description: "Safe, redacted diagnostic snapshot exposed outside Endpoint Platform."
      properties:
        schema_version:
          const: "endpoint_diagnostic_result_v1"
          title: "Schema Version"
          type: "string"
        profile:
          const: "diagnostic_v1"
          title: "Profile"
          type: "string"
        collected_at:
          format: "date-time"
          title: "Collected At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        reason:
          maxLength: 256
          minLength: 1
          pattern: "^(?![\\s\\S]*://)[^\\x00-\\x1f\\x7f]{1,256}(?![\\s\\S])"
          title: "Reason"
          type: "string"
        warnings:
          items:
            enum:
              - "command_failed"
              - "command_timed_out"
              - "data_truncated"
              - "permission_denied"
              - "probe_unavailable"
              - "redaction_applied"
              - "source_unavailable"
              - "unsupported_platform"
            type: "string"
          maxItems: 16
          title: "Warnings"
          type: "array"
        processes:
          items:
            "$ref": "#/components/schemas/EndpointDiagnosticResultV1__DiagnosticProcessV1"
          maxItems: 64
          title: "Processes"
          type: "array"
        log_excerpt:
          anyOf:
            -
              maxLength: 8192
              type: "string"
            -
              type: "null"
          default: null
          pattern: "^(?![\\s\\S]*(?:[Tt][Rr][Aa][Cc][Ee][Bb][Aa][Cc][Kk]\\s*\\(\\s*[Mm]ost\\s+[Rr]ecent\\s+[Cc]all\\s+[Ll]ast\\s*\\)\\s*:|[Ff]ile\\s+[\\\"'][^\\r\\n\\\"']+[\\\"']\\s*,\\s*[Ll]ine\\s+\\d+|[Bb][Ee][Aa][Rr][Ee][Rr]\\s+(?!(?:(?:[Aa]uth|[Aa]uthentication)\\s+(?:[Ww]as\\s+)?(?:\\[[Rr]edacted\\]|[Rr]edacted|[Aa]bsent|[Nn]one|[Nn]ull)|\\[[Rr]edacted\\]|[Rr]edacted|[Aa]bsent|[Nn]one|[Nn]ull)(?:[.,;)]?\\s*)$)\\S+|\\b(?:[Tt]oken|[Ss]ecret|[Aa][Cc][Cc][Ee][Ss][Ss][_ -]?[Tt][Oo][Kk][Ee][Nn]|[Aa][Pp][Ii][_ -]?[Kk][Ee][Yy]|[Cc][Ll][Ii][Ee][Nn][Tt][_ -]?[Ss][Ee][Cc][Rr][Ee][Tt]|[Pp][Aa][Ss][Ss][Ww][Oo][Rr][Dd]|[Cc][Rr][Ee][Dd][Ee][Nn][Tt][Ii][Aa][Ll]|[Aa][Uu][Tt][Hh][Oo][Rr][Ii][Zz][Aa][Tt][Ii][Oo][Nn])\\s*(?:[:=]|\\s+)\\s*(?!(?:\\[[Rr]edacted\\]|[Rr]edacted|[Aa]bsent|[Nn]one|[Nn]ull)(?:[.,;)]?\\s*)$)\\S+|(?:[Tt][Oo][Kk]|[Ss][Ee][Cc][Rr][Ee][Tt]|[Aa][Pp][Ii]|[Ss][Kk]|[Pp][Kk])_[A-Za-z0-9_-]{8,}|(?:[A-Za-z]:[\\\\/]|/[A-Za-z0-9._-]+(?:/[A-Za-z0-9._-]+)*)))[\\s\\S]{0,8192}(?![\\s\\S])"
          title: "Log Excerpt"
      required:
        - "schema_version"
        - "profile"
        - "collected_at"
        - "reason"
      title: "EndpointDiagnosticResultV1"
      type: "object"
    EndpointDiagnosticResultV1__DiagnosticProcessV1:
      additionalProperties: false
      properties:
        name:
          maxLength: 128
          minLength: 1
          title: "Name"
          type: "string"
        state:
          enum:
            - "running"
            - "sleeping"
            - "stopped"
            - "unknown"
          title: "State"
          type: "string"
      required:
        - "name"
        - "state"
      title: "DiagnosticProcessV1"
      type: "object"
...
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    OperationResponseData:
      properties:
        operation:
          "$ref": "#/components/schemas/EndpointOperationV1"
        result:
          anyOf:
            -
              "$ref": "#/components/schemas/EndpointDiagnosticResultV1"
            -
              type: "null"
      additionalProperties: false
      type: "object"
      required:
        - "operation"
        - "result"
      title: "OperationResponseData"
      description: "Public operation lifecycle and its optional validated safe result."
    OperationResponseEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/OperationResponseData"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "OperationResponseEnvelope"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
  securitySchemes:
    AgentBearer:
      type: "http"
      scheme: "bearer"
    ServiceBearer:
      type: "http"
      scheme: "bearer"